rename sliders to bars
