+++ src/components/ratings/RatingSummary/Bars/Bar/Bar.css
@@ -5,7 +5,7 @@
   margin-bottom: 1em;
 }
 
-.slider {
+.bar {
   display: flex;
   flex-direction: row;
   flex: 8;
