+++ src/components/ratings/RatingSummary/Bars/Bar/Bar.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import styles from './Slider.css';
+import styles from './Bar.css';
 
 export default function Slider({ stars, shaded }) {
   return (
     <div className={styles.container}>
       <div style={{ flex: 2 }}>{`${stars} stars:`}</div>
-      <div style={{ flex: 8 }} className={styles.slider}>
+      <div style={{ flex: 8 }} className={styles.bar}>
         <div style={{ flex: shaded }} className={styles.shaded} />
         <div style={{ flex: 1 - shaded }} className={styles.unshaded} />
       </div>
