+++ src/components/ratings/RatingSummary/Bars/Bars.css
