+++ src/components/ratings/RatingSummary/RatingSummary.jsx
@@ -2,14 +2,14 @@ import React from 'react';
 import styles from './RatingSummary.css';
 import Stars from './Stars/Stars';
 import Recommendations from './Recommendations/Recommendations';
-import Sliders from './Sliders/Sliders';
+import Bars from './Bars/Bars';
 
 export default function RatingSummary() {
   return (
     <div className={styles.container}>
       <Stars />
       <Recommendations />
-      <Sliders />
+      <Bars />
     </div>
   );
 }
